deploy/Daily-Summary/requirements.txt
mysql-connector-python==8.4.0
pydub==0.25.1
kokoro==0.9.4
soundfile==0.13.1
requests==2.32.3